fix(elk): Move elk log utils to utils_logging.py
