+++ astrosat/utils/utils_logging.py
@@ -1,6 +1,9 @@
 import logging
 import re
 import uuid
+import json
+from logstash.handler_tcp import TCPLogstashHandler
+from logstash.formatter import LogstashFormatterBase
 
 from astrosat.conf import app_settings as astrosat_settings
 
@@ -62,3 +65,52 @@ class DatabaseLogHandler(logging.Handler):
                 trace=trace,
             )
             db_record.tags.add(*tags)
+
+
+def format_elasticsearch_timestamp(time):
+    "Renders a timestamp in the format expected by elasticsearch"
+    return LogstashFormatterBase.format_timestamp(time)
+
+
+class ElasticsearchDocumentLogFormatter(LogstashFormatterBase):
+    """
+    Formats a json log record with additional fields needed for elasticsearch.
+    Input records MUST be JSON.
+    Based on LogstashFormatterBase/Version1 from python-logstash
+    https://github.com/vklochan/python-logstash/blob/master/logstash/formatter.py
+    """
+
+    def __init__(self, constant_fields=dict()):
+        self.constant_fields = constant_fields
+
+    def format(self, record):
+
+        record_fields = json.loads(record.getMessage())
+
+        meta_fields = {
+            "@timestamp": format_elasticsearch_timestamp(record.created),
+            "@version": 1,
+        }
+
+        # Merge analytics, constant field overrides and meta fields into one dict
+        logstash_message = {**record_fields, **self.constant_fields, **meta_fields}
+
+        return self.serialize(logstash_message)
+
+
+class AstrosatAppTCPLogstashLogHandler(TCPLogstashHandler):
+    """
+    Sends JSON log records to Logstash over TCP as line-seperated JSON documents
+    compatible with the Logstash TCP input plugin.
+    Adds extra metadata fields to each record about the app instance.
+    """
+
+    def __init__(self, host, port, app, instance, environment, stream="default"):
+        super(TCPLogstashHandler, self).__init__(host, port)
+
+        self.formatter = ElasticsearchDocumentLogFormatter(constant_fields={
+            "app": app,
+            "instance": instance,
+            "environment": environment,
+            "stream": stream,
+        })
